change datatype of table notifications

+++ db/migrations/000004_create_notifications_table.up.sql
@@ -1,9 +1,9 @@
 CREATE TABLE "notifications" (
     "id" bigserial PRIMARY KEY,
-    "user_id" int NOT NULL,
-    "status" int NOT NULL,
+    "user_id" bigint NOT NULL,
+    "status" bigint NOT NULL,
     "title" text NOT NULL,
     "body" text NOT NULL,
-    "is_read" int NOT NULL,
+    "is_read" bigint NOT NULL,
     "created_at" timestamp(0) NOT NULL
 )
